cleanup: Pass 3+4 — std::optional APIs, requires constraints, and bug fixes
Pass 3 — API modernization:
- bitword.hpp: 5 bool+output-ref methods → [[nodiscard]] std::optional<T>
  (get_next_set_bit, get_next_reset_bit, get_prev_set_bit, get_next_reset_bits,
   set_next_reset_bit); update all callers in bitset.hpp, compact_bitset.hpp,
   and test_bitword.cpp
- id_reserver.hpp: first_reserved_id / next_reserved_id → std::optional<uint32_t>;
  m_mutex upgraded to shared_mutex with shared_lock in read-only methods
- file_watcher.hpp/cpp: get_file_contents → std::optional<std::string>
- rpc_call.hpp: remove 6 #define type macros, replace with namespace-scope
  template alias templates and class-scope using aliases; enable_if → requires
- rpc_server.hpp: update signatures to use explicit template alias types

Pass 4 — bug fixes:
- obj_life_counter.hpp: copy ctor was incrementing s_alive but not s_created
- stream_tracker.hpp: move do_truncate from public to private
- metrics.hpp: m_Index field type size_t → uint64_t in Named{Counter,Gauge,Histogram}
  to match return type and numeric_limits<uint64_t>::max() sentinel

+++ include/sisl/metrics/metrics.hpp
@@ -170,7 +170,7 @@ private:
     NamedCounter() : m_Index{std::numeric_limits< uint64_t >::max()} {}
 
     static constexpr char m_Name[sizeof...(elements) + 1] = {elements..., '\0'};
-    size_t m_Index;
+    uint64_t m_Index;
 };
 
 template < char... elements >
@@ -194,7 +194,7 @@ private:
     NamedGauge() : m_Index{std::numeric_limits< uint64_t >::max()} {}
 
     static constexpr char m_Name[sizeof...(elements) + 1] = {elements..., '\0'};
-    size_t m_Index;
+    uint64_t m_Index;
 };
 
 template < char... elements >
@@ -218,7 +218,7 @@ private:
     NamedHistogram() : m_Index{std::numeric_limits< uint64_t >::max()} {}
 
     static constexpr char m_Name[sizeof...(elements) + 1] = {elements..., '\0'};
-    size_t m_Index;
+    uint64_t m_Index;
 };
 
 // decltype(BOOST_PP_CAT(BOOST_PP_STRINGIZE(name), _tstr)
